Split Thunk into its own file
